Clean up code, rebase on fnv-hash deletion
