+++ js/witness_calculator.js
@@ -17,7 +17,7 @@ limitations under the License.
 
 */
 
-import { flatArray, fnvHash } from "./utils.js";
+import { flatArray, fnvHash, fromArray32, toArray32 } from "./utils.js";
 import { Scalar, F1Field } from "ffjavascript";
 
 export default async function builder(code, options) {
@@ -454,32 +454,4 @@ class WitnessCalculatorCircom2 {
         return buff;
     }
 
-}
-
-
-function toArray32(s,size) {
-    const res = []; //new Uint32Array(size); //has no unshift
-    let rem = BigInt(s);
-    const radix = BigInt(0x100000000);
-    while (rem) {
-        res.unshift( Number(rem % radix));
-        rem = rem / radix;
-    }
-    if (size) {
-    var i = size - res.length;
-    while (i>0) {
-        res.unshift(0);
-        i--;
-    }
-    }
-    return res;
-}
-
-function fromArray32(arr) { //returns a BigInt
-    var res = BigInt(0);
-    const radix = BigInt(0x100000000);
-    for (let i = 0; i<arr.length; i++) {
-        res = res*radix + BigInt(arr[i]);
-    }
-    return res;
 }
